requirements.txt
python-dotenv==1.0.0
requests==2.31.0
ics==0.7.2
pytz==2023.3 